backend/requirements.txt
asgiref==3.11.1
certifi==2026.2.25
cfgv==3.5.0
charset-normalizer==3.4.7
distlib==0.4.0
Django==6.0.3
django-cors-headers==4.9.0
django-guardian==3.3.1
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
dotenv==0.9.9
filelock==3.25.2
fs-auth-middleware @ git+https://github.com/Cassio-de-Paula/FS_auth_middleware.git@9d4048bb6beb8887f649441d41d559eb5d605436
identify==2.6.18
idna==3.11
nodeenv==1.10.0
platformdirs==4.9.4
pre_commit==4.5.1
PyJWT==2.12.1
python-discovery==1.2.1
python-dotenv==1.2.2
PyYAML==6.0.3
requests==2.32.3
sqlparse==0.5.5
tzdata==2026.1
urllib3==2.6.3
virtualenv==21.2.0
